pyproject.toml
[tool.poetry]
name = "pybacked"
version = "0.1.0"
description = ""
authors = ["Vyacheslav Shuka <github@fwrdto.mailer.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.18.0"
pydantic = "^2.1.1"
click = "^8.1.6"
qrcode = "^7.4.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
